Rename SPFxAction.ts to SPFxActionBase.ts
