+++ apps/spfx-cli/src/cli/actions/CreateAction.ts
@@ -27,7 +27,7 @@ import {
 } from '@microsoft/spfx-template-api';
 
 import { SOLUTION_NAME_PATTERN } from '../../utilities/validation';
-import { SPFxActionBase } from './SPFxAction';
+import { SPFxActionBase } from './SPFxActionBase';
 
 // Deterministic namespace for CI mode GUIDs, derived from the well-known URL
 // namespace: uuidv5('spfx-cli:ci', '6ba7b810-9dad-11d1-80b4-00c04fd430c8')
